requirements.txt
langchain=0.1.16
langchain-google-genai=0.1.7
fastapi=0.109.0
uvicorn=0.23.2
google-generativeai=0.5.4
python-dotenv=1.0.1
requests=2.32.0
